# contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Decision": {
        "additionalProperties": false,
        "properties": {
          "agent_id": {
            "default": "",
            "title": "Agent Id",
            "type": "string"
          },
          "capability_result": {
            "additionalProperties": true,
            "title": "Capability Result",
            "type": "object"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionOutcome"
          },
          "execution_id": {
            "default": "",
            "title": "Execution Id",
            "type": "string"
          },
          "parameter_result": {
            "additionalProperties": true,
            "title": "Parameter Result",
            "type": "object"
          },
          "policy_results": {
            "additionalProperties": true,
            "title": "Policy Results",
            "type": "object"
          },
          "reasons": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Reason"
            },
            "title": "Reasons",
            "type": "array"
          },
          "risk": {
            "additionalProperties": true,
            "title": "Risk",
            "type": "object"
          },
          "task_id": {
            "default": "",
            "title": "Task Id",
            "type": "string"
          },
          "tool_name": {
            "default": "",
            "title": "Tool Name",
            "type": "string"
          },
          "trace_id": {
            "default": "",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "decision"
        ],
        "title": "Decision",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Phase 1 deterministic security Gateway.",
    "title": "AgentGuard Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/actions/evaluate": {
      "post": {
        "operationId": "evaluate_action_actions_evaluate_post",
        "parameters": [
          {
            "in": "query",
            "name": "track_request",
            "required": false,
            "schema": {
              "default": false,
              "title": "Track Request",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewayActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Decision"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Action",
        "tags": [
          "actions"
        ]
      }
    },
    "/actions/execute": {
      "post": {
        "operationId": "execute_action_actions_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewayActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Action",
        "tags": [
          "actions"
        ]
      }
    },
...
    "/agent/researcher/run": {
      "post": {
        "description": "Run the real Researcher Agent on a given task.\n\nThe agent uses the LLM to propose actions, sends each through the\nAgentGuard SDK \u2192 Gateway \u2192 Executor pipeline, and returns the\nfull trace of steps with their security decisions.",
        "operationId": "run_researcher_agent_agent_researcher_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Researcher Agent Agent Researcher Run Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Researcher Agent",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/approvals/{approval_id}/approve": {
      "post": {
        "operationId": "approve_action_approvals__approval_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Action Approvals  Approval Id  Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Action",
        "tags": [
          "approvals"
        ]
      }
    },
    "/approvals/{approval_id}/reject": {
      "post": {
        "operationId": "reject_action_approvals__approval_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reject Action Approvals  Approval Id  Reject Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Action",
        "tags": [
          "approvals"
        ]
      }
    },
    "/attack-lab/run/{scenario_id}": {
      "post": {
        "operationId": "run_attack_scenario_attack_lab_run__scenario_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Attack Scenario Attack Lab Run  Scenario Id  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Attack Scenario",
        "tags": [
          "attack-lab"
        ]
      }
    },
...
    "/demo/act1/legitimate": {
      "post": {
        "description": "ACT 1: Researcher agent reads a legitimate document \u2014 ALLOW.",
        "operationId": "demo_act1_legitimate_demo_act1_legitimate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoActRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Demo Act1 Legitimate Demo Act1 Legitimate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demo Act1 Legitimate",
        "tags": [
          "demo"
        ]
      }
    },
    "/demo/act2/prompt-injection": {
      "post": {
        "description": "ACT 2: Malicious document triggers prompt injection \u2192 AgentGuard BLOCKS.",
        "operationId": "demo_act2_prompt_injection_demo_act2_prompt_injection_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoActRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Demo Act2 Prompt Injection Demo Act2 Prompt Injection Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demo Act2 Prompt Injection",
        "tags": [
          "demo"
        ]
      }
    },
...
  }
}